Adding additions for debugging and cleaning a few things up.
FIXME:
- Currently there is a bug where the index calculated into
  the displacment map is beyond (or below) the size of the array.
  Committing this so it can be tested on another machine.

 NOTE: THIS IS BROKEN!

+++ medium.cpp
@@ -19,7 +19,7 @@ Medium::Medium()
 	this->initCommon();
 }
 
-Medium::Medium(const int x, const int y, const int z)
+Medium::Medium(const double x, const double y, const double z)
 {
 	this->z_bound = z;
 	this->y_bound = y;
